move checks to separate module
